Change the video loading and visualization technique (completely), Fix bugs
Change the video loading and visualization technique (completely): now videos are not completely loaded in RAM, but a frame is seeked at each update of current frame number. A step in seconds is required (in object `config`). This is a breaking change, previous annotation may not be aligned to the video anymore.
Fix bugs with importJson function and (unexpected) warning alert on json download
Add visualization of current frame number